pyproject.toml
[project]
name = "carla-gym"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "addict==2.4.0",
    "antlr4-python3-runtime==4.8",
    "carla==0.9.15",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "cloudpickle==3.1.1",
    "cycler==0.12.1",
    "decorator==4.4.2",
    "dictor==0.1.7",
    "docker-pycreds==0.4.0",
    "Farama-Notifications==0.0.4",
    "filelock==3.16.1",
    "fsspec==2025.2.0",
    "gitdb==4.0.12",
    "GitPython==3.1.44",
    "gql==0.2.0",
    "graphql-core==1.1",
    "gym-notices==0.0.8",
    "gymnasium==1.1.0",
    "h5py==3.11.0",
    "hf-xet==1.1.3",
    "huggingface-hub==0.32.4",
    "hydra-core==1.0.3",
    "idna==3.10",
    "imageio==2.35.1",
    "imageio-ffmpeg==0.5.1",
    "imgaug==0.4.0",
    "importlib-resources==3.0.0",
    "importlib_metadata==8.5.0",
    "Jinja2==3.1.5",
    "kiwisolver==1.4.7",
    "MarkupSafe==2.1.5",
    "matplotlib>=3.5.0",
    "moviepy==1.0.3",
    "mpmath==1.3.0",
    "networkx==3.1",
    "numpy==1.24.4",
    "nvidia-cublas-cu12==12.1.3.1",
    "nvidia-cuda-cupti-cu12==12.1.105",
    "nvidia-cuda-nvrtc-cu12==12.1.105",
    "nvidia-cuda-runtime-cu12==12.1.105",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.0.2.54",
    "nvidia-curand-cu12==10.3.2.106",
    "nvidia-cusolver-cu12==11.4.5.107",
    "nvidia-cusparse-cu12==12.1.0.106",
    "nvidia-nccl-cu12==2.20.5",
    "nvidia-nvjitlink-cu12==12.8.61",
    "nvidia-nvtx-cu12==12.1.105",
    "omegaconf==2.1.2",
    "opencv-python>=4.6.0",
    "packaging==25.0",
    "pandas==2.0.3",
    "pillow==10.4.0",
    "platformdirs==4.3.6",
    "proglog==0.1.12",
    "promise==2.3",
    "protobuf==5.29.3",
    "psutil==6.1.1",
    "py-trees==0.8.3",
    "pydot==1.4.1",
    "pygame==2.6.1",
    "pygccxml==2.0.1",
    "pyparsing==3.1.4",
    "pyplusplus==1.8.4",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "PyWavelets>=1.4.0",
    "PyYAML==6.0.2",
    "requests==2.32.3",
    "safetensors==0.5.3",
    "scikit-image>=0.18.0",
    "scipy==1.10.1",
    "sentry-sdk==2.20.0",
    "setproctitle==1.3.4",
    "shapely==2.0.7",
    "Shimmy==2.0.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "stable_baselines3==2.4.1",
    "subprocess32==3.5.4",
    "sympy==1.13.3",
    "tifffile>=2021.1.0",
    "timm==1.0.15",
    "torch==2.4.1",
    "torchvision==0.19.1",
    "tqdm==4.67.1",
    "triton==3.0.0",
    "typing_extensions>=4.14.0",
    "tzdata==2025.1",
    "urllib3==2.2.3",
    "wandb==0.18.3",
    "zipp==3.20.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["carla_gym*"]

[tool.ruff]
exclude = [".git", ".mypy_cache", ".ruff_cache", ".venv", "third_party"]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules/
fixable = ["ALL"]
unfixable = []
select = [
    "A",  # flake8-builtin
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W",  # pycodestyle warning
    "PL",  # Pylint
    "UP",  # pyupgrade
]
ignore = [
    "B905",  # Zip-without-explicit-strict
    "E501",  # Line too long
    "F403",  # Unable to detect undefined names
    "F722",  # Checks for forward annotations
    "F821",  # Checks for uses of undefined names
    "N806",  # non-lowercase-variable-in-function
    "N812",  # Lowercase imported as non-lowercase
    "N999",  # Invalid module name
    "PLR0912",  # Too many branches
    "PLR0913",  # Too many arguments in function definition
    "PLR2004",  # Magic value used in comparison
    "PLW2901",  # redefined-loop-name
]

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
